src/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5.3
colorama==0.4.3
distlib==0.3.0
Django==3.0.4
django-crispy-forms==1.9.1
django-grappelli==2.14.2
djangorestframework==3.11.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.4.4
pysciter==0.4.30
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
virtualenv==20.0.17
wrapt==1.11.2
